feat(auth): added protected routes
protected routes for authenticated users

BREAKING CHANGE: This change will enable protected routes for the application, meaning now we need
to be logged in as either "member" or an "admin" user. /dashboard is for admin where /home is for
member. if no user is logged in then app is redirected to /landing for signin or signup.

fix #17